+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 4.1.6
 - Fixed GenshinData repo id for GitLab API again.
+- Updated the branch name of GenshinData GitLab repo.
 # 4.1.5
 - Removed WeaponRefinements.
 - Fixed GenshinData repo id for GitLab API.